platform: nordic_nrf: Make tfm_hal_platform_init overridable
In TF-M 1.3 the tfm_hal_platform_init function was declared as WEAK,
making it possible for out-of-tree platforms to add init handling
that could only be done out-of-tree, e.g. init of libraries.

Re-enable the tfm_hal_platform_init to be overridable as well as
re-using the common handling by moving it to the board specific
part of the platform-support, and calling a common init function.

Change-Id: Id334e35a8fec8764fe00cc00e60ba945e0098da7
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>