platform: nordic_nrf: Fix build when SECURE_UART1 has been disabled
Fix build when SECURE_UART1 has been disabled. We had excluded the UART
driver in this configuration, but this is not correct as secure log
levels needs to be set to silence in order to not have any stdout
calls.
This fails at link-time because there are still references to the UART
driver struct.
Use the TFM_SPM_LOG_RAW_ENABLED config instead to exclude the UART
driver for the nRF platform when logging is not enabled.

Change-Id: I6e2990e1a0bf5e1d236c4fb4bb258d3e5a555fc1
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>