platform: link platform_s library against mbedcrypto
Link the platform_s library against the mbedcrypto library.
The link against mbedtls for crypto_keys is incorrect since crypto_key
relies on mbedtls/hkdf.h header, which is part of the mbedcrypto library
exported by mbedtls project.
Reduce the conditional to crypto partition enabled since there are many
other cases where the mbedcrypto library is a dependency, such as
crypto_nv_seed and the accelerators.
The link is needed in order to propagate interface include and compile
options to the platform_s source files.

Change-Id: Ice918034997ae3fad7e17c3bcdc87abda43bb858
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>